+++ bin/genotype.py
