+++ nextflow.config
@@ -1,7 +1,7 @@
 manifest {
   author = 'Sherrie Wang (nextflow pipeline)/Kevin Kuchinski (FluViewer), adapted by Sherrie Wang'
   name = 'BCCDC-PHL/hcv-nf'
-  version = '0.1.0'
+  version = '1.0.1'
   description = 'BCCDC-PHL HCV genotyping pipeline'
   mainScript = 'main.nf'
   nextflowVersion = '>=20.01.0'
